api/requirements.txt
Django==4.1
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-gis==1.0
gunicorn==20.1.0
psycopg2==2.9.3
